pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "applied-image-analyzer"
version = "0.1.0"
dependencies = [
  "streamlit",
  "pandas",
  "pillow",
  "numpy",
  "scikit-image"
]